ImageOptimizer.app/requirements.txt
# Image Optimizer backend ? only packages used by this app
Flask>=2.3
flask-cors>=4.0
opencv-python>=4.8
Pillow>=10.0
